add no entity event

+++ src/main/kotlin/br/com/invite/repository/UserRepository.kt
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository: JpaRepository<User, Long> {
 
-    fun countUserByName
-            (name: Name): Int
+    fun countUserByName(name: Name): Long
 
 }
